Fix bug when printing more than one symbol with printf
e.g. The first symbol would be blank/corrupted when running:
    printf("%s %s\n", func, func)

This was caused by a dodgy conversion from std::string
to C-string, where the result of ".c_str()" was being
saved, but the underlying std::string was modified by
the vector it was being stored in, invalidating the C-string

+++ src/bpftrace.h
@@ -25,8 +25,8 @@ public:
   int run();
   int print_maps();
   std::string get_stack(uint32_t stackid, bool ustack, int indent=0);
-  std::string resolve_sym(uint64_t addr, bool show_offset=false);
-  std::string resolve_usym(uint64_t addr) const;
+  std::string resolve_sym(uintptr_t addr, bool show_offset=false);
+  std::string resolve_usym(uintptr_t addr) const;
 
   std::map<std::string, std::unique_ptr<IMap>> maps_;
   std::map<std::string, std::tuple<uint8_t *, uintptr_t>> sections_;
